requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-cacheops==4.0.6
djangorestframework==3.8.2
drf-nested-routers==0.90.2
funcy==1.10.1
httpie==0.9.9
idna==2.6
jsonfield==2.0.2
pipenv==2018.5.18
Pygments==2.2.0
pytz==2018.4
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0